Cleanup cmake file. Add another static lib that we link the lib against. This can then also link the tests
